:beetle: Fix model check errors.
./manage.py test

SystemCheckError: System check identified some issues:

ERRORS:
apps.CrushOrder.vessels: (fields.E300) Field defines a relation with model 'apps.Vessel', which is either not installed, or is abstract.
apps.CrushOrderVesselMapping.vessel: (fields.E300) Field defines a relation with model 'apps.Vessel', which is either not installed, or is abstract.
