+++ apps/serializers.py
@@ -4,7 +4,7 @@ from apps.models.crush_orders import CrushOrder
 from apps.models.dips import Dip
 from apps.models.dockets import Docket
 from apps.models.fruit_intakes import FruitIntake
-from apps.models.models import CrushOrderDocketMapping
+from apps.models import CrushOrderDocketMapping
 from apps.models.vessels import Vessel
